+++ lib/modugo.dart
@@ -5,8 +5,8 @@ export 'src/module.dart';
 export 'src/manager.dart';
 export 'src/transition.dart';
 
-export 'src/events/event_module.dart';
 export 'src/events/event_channel.dart';
+export 'src/events/event_registry.dart';
 
 export 'src/widgets/modugo_loader_widget.dart';
