+++ test/events/event_registry_test.dart
@@ -1,32 +1,27 @@
 import 'dart:async';
 
-import 'package:event_bus/event_bus.dart';
+import 'package:modugo/src/module.dart';
 import 'package:flutter_test/flutter_test.dart';
 
-import 'package:modugo/src/events/event_module.dart';
 import 'package:modugo/src/events/event_channel.dart';
+import 'package:modugo/src/events/event_registry.dart';
 
 void main() {
-  group('EventModule Tests', () {
-    late EventBus eventBus;
+  group('EventRegistry Tests', () {
     late _EventModule module;
 
     setUp(() {
-      eventBus = EventBus();
-      module = _EventModule(eventBus: eventBus);
+      module = _EventModule();
     });
 
     test('initState calls listen', () async {
       final completer = Completer<void>();
 
-      final module2 = _EventModule(
-        eventBus: eventBus,
-        onEventCalled: (_) => completer.complete(),
-      );
+      final module2 = _EventModule(onEventCalled: (_) => completer.complete());
 
       module2.initState();
 
-      EventChannel.emit(_EventMock('Hello'), eventBus: eventBus);
+      EventChannel.emit(_EventMock('Hello'));
 
       await completer.future;
 
@@ -41,7 +36,7 @@ void main() {
         completer.complete();
       });
 
-      EventChannel.emit(_EventMock('Test'), eventBus: eventBus);
+      EventChannel.emit(_EventMock('Test'));
 
       await completer.future;
 
@@ -55,7 +50,7 @@ void main() {
 
       module.dispose();
 
-      EventChannel.emit(_EventMock('Test'), eventBus: eventBus);
+      EventChannel.emit(_EventMock('Test'));
 
       await Future.delayed(Duration.zero);
 
@@ -69,7 +64,7 @@ void main() {
 
       module.dispose();
 
-      EventChannel.emit(_EventMock('Test'), eventBus: eventBus);
+      EventChannel.emit(_EventMock('Test'));
 
       await Future.delayed(Duration.zero);
       expect(callbackCalled, false);
@@ -82,10 +77,10 @@ final class _EventMock {
   _EventMock(this.message);
 }
 
-final class _EventModule extends EventModule {
+final class _EventModule extends Module with EventRegistry {
   final void Function(_EventMock event)? onEventCalled;
 
-  _EventModule({super.eventBus, this.onEventCalled});
+  _EventModule({this.onEventCalled});
 
   @override
   void listen() {
